requirements-api.txt
# =============================================================================
# CSM Voice Service - API Dependencies
# =============================================================================
#
# These are additional dependencies for the production API service.
# Install alongside the base requirements.txt:
#   pip install -r requirements.txt -r requirements-api.txt
#

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Async support
httpx>=0.25.0
websockets>=12.0
aiofiles>=23.2.0

# Database (PostgreSQL)
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.12.0

# Caching
redis>=5.0.0
cachetools>=5.3.0

# Audio processing (in addition to torchaudio)
soundfile>=0.12.0
scipy>=1.11.0
numpy>=1.24.0

# Observability
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-exporter-otlp>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0
prometheus-client>=0.18.0

# Security
python-jose[cryptography]>=3.3.0
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Rate limiting
slowapi>=0.1.9

# Cloud providers (for training)
# runpod>=1.4.0  # Uncomment for RunPod support
# modal>=0.55.0  # Uncomment for Modal support

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0

# Code quality
black>=23.10.0
ruff>=0.1.0
mypy>=1.6.0
